// apps/chat/utils/types.ts
export type CreateGroupParams = {
  creator: string;
  title?: string;
  users: string[];
};
// ...
export type AccessParams = {
  id: number;
  userId: number;
};

export type TransferOwnerParams = {
  userId: number;
  groupId: number;
  newOwnerId: number;
};
// ...
export interface Attachment extends Express.Multer.File {}
// ...
export type UpdateGroupDetailsParams = {
  id: number;
  title?: string;
  avatar?: Attachment;
};